+++ Sources/RichTextKit/Images/RichTextImageAttachmentManager.swift
@@ -21,8 +21,8 @@ import AppKit
  This protocol extends ``RichTextReader`` with functionality
  for handling image attachments.
 
- The protocol is implemented by `NSAttributedString` as well
- as other types in the library.
+ The protocol is implemented by `NSAttributedString` but can
+ be oimplemented by any `RichTextReader`.
  */
 public protocol RichTextImageAttachmentManager: RichTextReader {}
 
@@ -35,7 +35,8 @@ public extension RichTextImageAttachmentManager {
      */
     func attachmentBounds(
         for image: ImageRepresentable,
-        maxSize: CGSize) -> CGRect {
+        maxSize: CGSize
+    ) -> CGRect {
         let size = attachmentSize(for: image, maxSize: maxSize)
         return CGRect(origin: .zero, size: size)
     }
@@ -45,7 +46,8 @@ public extension RichTextImageAttachmentManager {
      */
     func attachmentSize(
         for image: ImageRepresentable,
-        maxSize: CGSize) -> CGSize {
+        maxSize: CGSize
+    ) -> CGSize {
         let size = image.size
         let validWidth = size.width < maxSize.width
         let validHeight = size.height < maxSize.height
@@ -79,15 +81,6 @@ public extension RichTextImageAttachmentManager {
             if oldBounds == newBounds { return }
             attachment.bounds = newBounds
         }
-    }
-}
-
-private extension NSTextAttachment {
-
-    var attachedImage: ImageRepresentable? {
-        if let image = self.image { return image }
-        guard let imageData = self.contents else { return nil }
-        return ImageRepresentable(data: imageData)
     }
 }
 #endif