+++ Sources/RichTextKit/RichTextEditor.swift
@@ -103,7 +103,10 @@ public struct RichTextEditor: ViewRepresentable {
         return textView
     }
 
-    public func updateUIView(_ view: UIViewType, context: Context) {}
+    public func updateUIView(_ view: UIViewType, context: Context) {
+        textView.attributedString = text.wrappedValue
+        richTextContext.highlightedRange = nil
+    }
     #endif
 
     #if os(macOS)
@@ -113,7 +116,10 @@ public struct RichTextEditor: ViewRepresentable {
         return scrollView
     }
 
-    public func updateNSView(_ view: NSViewType, context: Context) {}
+    public func updateNSView(_ view: NSViewType, context: Context) {
+        textView.attributedString = text.wrappedValue
+        richTextContext.highlightedRange = nil
+    }
     #endif
 }