+++ Sources/RichTextKit/RichTextView_UIKit.swift
@@ -166,15 +166,15 @@ open class RichTextView: UITextView, RichTextViewComponent {
     ) {
         attributedString = .empty
         setupInitialFontSize()
+        imageConfiguration = standardImageConfiguration(for: format)
+        text.autosizeImageAttachments(maxSize: imageAttachmentMaxSize)
         attributedString = text
         allowsEditingTextAttributes = false
         autocapitalizationType = .sentences
         backgroundColor = .clear
-        imageConfiguration = standardImageConfiguration(for: format)
-        text.autosizeImageAttachments(maxSize: imageAttachmentMaxSize)
         richTextDataFormat = format
         spellCheckingType = .no
-        if text.string.isEmpty {
+        trySetupInitialTextColor(for: text) {
             textColor = .label
         }
         setContentCompressionResistancePriority(.defaultLow, for: .horizontal)